Adding snack bar, V 1.1

+++ todo-list/src/Components/Tasks.jsx
@@ -1,51 +1,33 @@
 import DoneIcon from "@mui/icons-material/Done";
 import CreateIcon from "@mui/icons-material/Create";
 import DeleteIcon from "@mui/icons-material/Delete";
-import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  IconButton,
-  TextField,
-  ToggleButton,
-  ToggleButtonGroup,
-} from "@mui/material";
+import { IconButton, ToggleButton, ToggleButtonGroup } from "@mui/material";
 import { TasksContext } from "./TasksContext";
-import { useContext, useState } from "react";
-export default function Tasks() {
+import { SnackContext } from "./SnackContext";
+import { useContext, useState, useMemo } from "react";
+export default function Tasks({ openDialog }) {
   const { tasksData, setTasksData } = useContext(TasksContext);
-  const [editedTask, setEditedTask] = useState("");
-  const [open, setOpen] = useState(false);
+  const { handleShowHide, setSnackMsg } = useContext(SnackContext);
   const [filtredTasks, setFiltredTasks] = useState("all");
 
-  const handleClose = () => {
-    setOpen(false);
-  };
-
   let tasksCompletion = tasksData;
-  const unCompletedTasks = tasksData.filter((t) => !t.isDone);
+  const unCompletedTasks = useMemo(() => {
+    return tasksData.filter((t) => !t.isDone);
+  }, [tasksData]);
 
-  const completedTasks = tasksData.filter((t) => t.isDone);
+  const completedTasks = useMemo(() => {
+    return tasksData.filter((t) => t.isDone);
+  }, [tasksData]);
 
-  if (filtredTasks == "completed") {
+  if (filtredTasks === "completed") {
     tasksCompletion = completedTasks;
-  } else if (filtredTasks == "uncompleted") {
+  } else if (filtredTasks === "uncompleted") {
     tasksCompletion = unCompletedTasks;
   }
+
   const tasksList = tasksCompletion.map((task) => {
-    function handleSubmitBtn() {
-      const updateTask = tasksData.map((t) => {
-        if (t.id === task.id) {
-          return { ...t, title: editedTask };
-        } else {
-          return t;
-        }
-      });
-      setTasksData(updateTask);
-      localStorage.setItem("task", JSON.stringify(updateTask));
-      handleClose();
+    function handleOpenDialog() {
+      openDialog(task);
     }
     return (
       <div
@@ -58,44 +40,9 @@ export default function Tasks() {
           textDecoration: task.isDone ? "line-through" : "none",
         }}
       >
-        <Dialog open={open} onClose={handleClose}>
-          <DialogTitle>Edit Your Task</DialogTitle>
-          <DialogContent>
-            <TextField
-              autoFocus
-              required
-              margin="dense"
-              id="name"
-              label="Task"
-              fullWidth
-              variant="standard"
-              value={editedTask}
-              onChange={(e) => {
-                setEditedTask(e.target.value);
-              }}
-            />
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={handleClose} variant="outlined" color="error">
-              Cancel
-            </Button>
-            <Button
-              onClick={handleSubmitBtn}
-              variant="outlined"
-              color="primary"
-            >
-              Submit
-            </Button>
-          </DialogActions>
-        </Dialog>
         <p>{task.title}</p>
         <div className="cta-btns">
-          <IconButton
-            color="primary"
-            onClick={() => {
-              setOpen(true);
-            }}
-          >
+          <IconButton color="primary" onClick={handleOpenDialog}>
             <CreateIcon />
           </IconButton>
           <IconButton
@@ -104,6 +51,10 @@ export default function Tasks() {
               const checkIsDone = tasksData.map((t) => {
                 if (t.id === task.id) {
                   t.isDone = !t.isDone;
+                  if (t.isDone) {
+                    handleShowHide();
+                    setSnackMsg("Your task is completed!");
+                  }
                 }
                 return t;
               });
@@ -116,15 +67,21 @@ export default function Tasks() {
           <IconButton
             color="error"
             onClick={() => {
-              const deleteTask = tasksData.filter((t) => {
-                if (t.id === task.id) {
-                  return t.id !== task.id;
-                }
-                return t;
-              });
-
-              setTasksData(deleteTask);
-              localStorage.setItem("task", JSON.stringify(deleteTask));
+              const confirmToDelete = window.confirm(
+                "I confirm to delete the task"
+              );
+              if (confirmToDelete) {
+                handleShowHide();
+                setSnackMsg("Your task is removed!");
+                const deleteTask = tasksData.filter((t) => {
+                  if (t.id === task.id) {
+                    return t.id !== task.id;
+                  }
+                  return t;
+                });
+                setTasksData(deleteTask);
+                localStorage.setItem("task", JSON.stringify(deleteTask));
+              }
             }}
           >
             <DeleteIcon />
@@ -133,9 +90,12 @@ export default function Tasks() {
       </div>
     );
   });
+
   return (
     <>
-      <div style={{ margin: "1rem" }}>
+      <div
+        style={{ margin: "1rem", display: "flex", justifyContent: "center" }}
+      >
         <ToggleButtonGroup
           value={filtredTasks}
           exclusive
